Split gmt_offset, time and tzname searches to its own separate strategy under the advanced package and remove them from match
